feat(sdk): remove warrant dependency to remove native dependencies
The SDK currently uses a small part of the https://github.com/capless/warrant library to hand Secure Remote Password (http://srp.stanford.edu/) however this library also provides a number of utility functions for cognito which we don't use.

This PR takes just the code we use from warrant and copies it into this SDK, with a header including reference to the library and it's author.

This change enables customers to install the SDK on OSX and upload it to lambda and run without the need to recompile native deps in linux containers.

fixes #60